DataServer moved to dedicated module
