+++ utils/__init__.py
@@ -1,4 +1,5 @@
 from .config import Config
 from .data_file import DataFile
+from .data_server import DataServer
 from .wifi import WiFi
 from .api_calls import *
